Add bottom tabs with projects and contact
